requirements.txt
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
supabase==2.7.4
openai==1.35.15
python-dotenv==1.0.1
email-validator==2.2.0